requirements.txt
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
apispec==6.7.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.1
marshmallow==3.23.1
openai==1.51.2
packaging==24.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
webargs==8.6.0
Werkzeug==3.0.4
WTForms==3.2.1